Rename budget to law of motion
